+++ cmd/complete_cmd.go
@@ -25,13 +25,12 @@ import (
 )
 
 type CompleteCmd struct {
-	Install   bool `kong:"cmd,short='I',help='Install shell completions',xor='action'"`
-	Uninstall bool `kong:"cmd,short='U',help='Uninstall shell completions',xor='action'"`
-	Pre19     bool `kong:"cmd,help='Modify older pre-v1.9 shell completion integration'"`
+	Install   bool `kong:"short='I',help='Install shell completions',xor='action'"`
+	Uninstall bool `kong:"short='U',help='Uninstall shell completions',xor='action'"`
+	Pre19     bool `kong:"help='Modify older pre-v1.9 shell completion integration'"`
 }
 
 func (cc *CompleteCmd) Run(ctx *RunContext) error {
-
 	if ctx.Cli.Completions.Install {
 		if ctx.Cli.Completions.Pre19 {
 			kp := &kongplete.InstallCompletions{}