+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+### New Features 
+
+ * Add support for `aws-sso-profile` and `aws-sso-clear` helper functions which 
+	are automatically installed #357
+ * Add `completions` command which supports `--install` and `--uninstall` flags
+
+### Changes
+
+ * `install-completions` is no longer a valid command.
+
 ## [v1.8.1] - 2022-05-02
 
 ### New Features